refactor move Modals.tsx to modals/ directory
